Initial work on conversion from bioconda to biocontainers
